refactor(ui): split skill workshop state
