+++ ui/src/pages/skill-workshop/proposals.ts
@@ -9,12 +9,18 @@ import {
 } from "../../lib/sessions/session-key.ts";
 import type {
   SkillWorkshopAction,
-  SkillWorkshopActionNotice,
-  SkillWorkshopMode,
   SkillWorkshopProposal,
   SkillWorkshopProposalStatus,
-  SkillWorkshopStatusFilter,
 } from "../../lib/skill-workshop/index.ts";
+import { createSkillWorkshopHistoryScanState, type SkillWorkshopState } from "./state.ts";
+export {
+  createSkillWorkshopState,
+  skillWorkshopRouteData,
+  type SkillWorkshopHistoryScanResult,
+  type SkillWorkshopHistoryScanState,
+  type SkillWorkshopRouteData,
+  type SkillWorkshopState,
+} from "./state.ts";
 
 const SKILL_WORKSHOP_NOTICE_MS = 2800;
 
@@ -40,37 +46,6 @@ type SkillProposalManifest = {
   updatedAt: string;
   proposals: SkillProposalManifestEntry[];
 };
-
-export type SkillWorkshopHistoryScanResult = {
-  schema: "openclaw.skill-workshop.history-scan.v1";
-  hasScanned: boolean;
-  reviewedSessions: number;
-  ideasFound: number;
-  hasMore: boolean;
-  lastScanReviewed: number;
-  lastScanIdeas: number;
-  lastScanAt?: string;
-  oldestReviewedAt?: string;
-  newestReviewedAt?: string;
-};
-
-export type SkillWorkshopHistoryScanState = {
-  loading: boolean;
-  loaded: boolean;
-  running: boolean;
-  error: string | null;
-  result: SkillWorkshopHistoryScanResult | null;
-};
-
-function createSkillWorkshopHistoryScanState(): SkillWorkshopHistoryScanState {
-  return {
-    loading: false,
-    loaded: false,
-    running: false,
-    error: null,
-    result: null,
-  };
-}
 
 type SkillProposalSupportFileRecord = {
   path: string;
@@ -116,88 +91,6 @@ export type SkillWorkshopContext = {
   gateway: ApplicationGateway;
   agentSelection: Pick<AgentSelectionCapability, "state">;
 };
-
-export type SkillWorkshopState = {
-  skillWorkshopAgentId: string | null;
-  skillWorkshopLoading: boolean;
-  skillWorkshopLoaded: boolean;
-  skillWorkshopError: string | null;
-  skillWorkshopInspectingKey: string | null;
-  skillWorkshopProposals: SkillWorkshopProposal[];
-  skillWorkshopSelectedKey: string | null;
-  skillWorkshopActionBusy: { key: string; action: SkillWorkshopAction } | null;
-  skillWorkshopActionNotice: SkillWorkshopActionNotice | null;
-  skillWorkshopActionNoticeTimer?: ReturnType<typeof globalThis.setTimeout> | number | null;
-  skillWorkshopRevisionKey: string | null;
-  skillWorkshopRevisionDraft: string;
-  skillWorkshopStatusFilter: SkillWorkshopStatusFilter;
-  skillWorkshopQuery: string;
-  skillWorkshopFilePreviewKey: string | null;
-  skillWorkshopFilePreviewQuery: string;
-  skillWorkshopQueueWidth: number;
-  skillWorkshopMode: SkillWorkshopMode;
-  skillWorkshopUseCurrentChatForRevisions: boolean;
-  skillWorkshopHistoryScan: SkillWorkshopHistoryScanState;
-};
-
-export type SkillWorkshopRouteData = Pick<
-  SkillWorkshopState,
-  | "skillWorkshopAgentId"
-  | "skillWorkshopLoading"
-  | "skillWorkshopLoaded"
-  | "skillWorkshopError"
-  | "skillWorkshopInspectingKey"
-  | "skillWorkshopProposals"
-  | "skillWorkshopSelectedKey"
-  | "skillWorkshopActionBusy"
-  | "skillWorkshopActionNotice"
-  | "skillWorkshopRevisionKey"
-  | "skillWorkshopRevisionDraft"
-  | "skillWorkshopHistoryScan"
->;
-
-export function createSkillWorkshopState(data?: SkillWorkshopRouteData): SkillWorkshopState {
-  return {
-    skillWorkshopAgentId: data?.skillWorkshopAgentId ?? null,
-    skillWorkshopLoading: data?.skillWorkshopLoading ?? false,
-    skillWorkshopLoaded: data?.skillWorkshopLoaded ?? false,
-    skillWorkshopError: data?.skillWorkshopError ?? null,
-    skillWorkshopInspectingKey: data?.skillWorkshopInspectingKey ?? null,
-    skillWorkshopProposals: data?.skillWorkshopProposals ?? [],
-    skillWorkshopSelectedKey: data?.skillWorkshopSelectedKey ?? null,
-    skillWorkshopActionBusy: data?.skillWorkshopActionBusy ?? null,
-    skillWorkshopActionNotice: data?.skillWorkshopActionNotice ?? null,
-    skillWorkshopActionNoticeTimer: null,
-    skillWorkshopRevisionKey: data?.skillWorkshopRevisionKey ?? null,
-    skillWorkshopRevisionDraft: data?.skillWorkshopRevisionDraft ?? "",
-    skillWorkshopStatusFilter: "pending",
-    skillWorkshopQuery: "",
-    skillWorkshopFilePreviewKey: null,
-    skillWorkshopFilePreviewQuery: "",
-    skillWorkshopQueueWidth: 360,
-    skillWorkshopMode: "today",
-    skillWorkshopUseCurrentChatForRevisions: false,
-    skillWorkshopHistoryScan:
-      data?.skillWorkshopHistoryScan ?? createSkillWorkshopHistoryScanState(),
-  };
-}
-
-export function skillWorkshopRouteData(state: SkillWorkshopState): SkillWorkshopRouteData {
-  return {
-    skillWorkshopAgentId: state.skillWorkshopAgentId,
-    skillWorkshopLoading: state.skillWorkshopLoading,
-    skillWorkshopLoaded: state.skillWorkshopLoaded,
-    skillWorkshopError: state.skillWorkshopError,
-    skillWorkshopInspectingKey: state.skillWorkshopInspectingKey,
-    skillWorkshopProposals: state.skillWorkshopProposals,
-    skillWorkshopSelectedKey: state.skillWorkshopSelectedKey,
-    skillWorkshopActionBusy: state.skillWorkshopActionBusy,
-    skillWorkshopActionNotice: state.skillWorkshopActionNotice,
-    skillWorkshopRevisionKey: state.skillWorkshopRevisionKey,
-    skillWorkshopRevisionDraft: state.skillWorkshopRevisionDraft,
-    skillWorkshopHistoryScan: state.skillWorkshopHistoryScan,
-  };
-}
 
 function getErrorMessage(err: unknown): string {
   return err instanceof Error ? err.message : String(err);
